pre-commit fixes and linting

+++ crystal_toolkit/components/diffraction.py
@@ -375,9 +375,9 @@ crystals in a spherical shape is used. However, in practice K can vary from 0.62
                 )
                 sigma = (alpha / np.sqrt(2 * np.log(2))).item()
 
-                center_idx = int(round((xp - first) * N_density))
+                center_idx = round((xp - first) * N_density)
                 # total broadening window of 2 * num_sigma
-                half_window = int(round(num_sigma * sigma * N_density))
+                half_window = round(num_sigma * sigma * N_density)
 
                 lb = max(0, (center_idx - half_window))
                 ub = min(N, (center_idx + half_window))
